Adding abstract interactor
