+++ src/Services/Identity/Command/Application/Abstractions/Gateways/IEventStoreGateway.cs
@@ -5,6 +5,7 @@ namespace Application.Abstractions.Gateways;
 public interface IEventStoreGateway
 {
     Task AppendAsync(IAggregateRoot aggregate, CancellationToken cancellationToken);
+
     Task<IAggregateRoot> LoadAsync<TAggregate>(Guid aggregateId, CancellationToken cancellationToken)
         where TAggregate : IAggregateRoot, new();
 }
