+++ src/Services/Identity/Command/Application/Abstractions/Interactors/IInteractor.cs
@@ -7,15 +7,3 @@ public interface IInteractor<in TMessage>
 {
     Task InteractAsync(TMessage message, CancellationToken cancellationToken);
 }
-
-public interface ICommandInteractor<in TCommand> : IInteractor<TCommand>
-    where TCommand : ICommandWithId
-{
-    new Task InteractAsync(TCommand command, CancellationToken cancellationToken);
-}
-
-public interface IEventInteractor<in TEvent> : IInteractor<TEvent>
-    where TEvent : IEventWithId
-{
-    new Task InteractAsync(TEvent @event, CancellationToken cancellationToken);
-}
