+++ src/Services/Identity/Command/Application/UseCases/DeactiveteUserInteractor.cs
@@ -1,17 +1,17 @@
 using Application.Abstractions;
 using Application.Abstractions.Gateways;
 using Application.Abstractions.Interactors;
-using Contracts.Services.Account;
+using Contracts.Services.Identity;
 using Domain.Aggregates;
-using Command = Contracts.Services.Identity.Command;
+using DomainEvent = Contracts.Services.Account.DomainEvent;
 
 namespace Application.UseCases;
 
 public class DeactivateUserInteractor : EventInteractor<User, DomainEvent.AccountDeactivated>
 {
-    public DeactivateUserInteractor(IEventStoreGateway eventStoreGateway, IEventBusGateway eventBusGateway, IUnitOfWork unitOfWork) 
+    public DeactivateUserInteractor(IEventStoreGateway eventStoreGateway, IEventBusGateway eventBusGateway, IUnitOfWork unitOfWork)
         : base(eventStoreGateway, eventBusGateway, unitOfWork) { }
-    
+
     public override Task InteractAsync(DomainEvent.AccountDeactivated @event, CancellationToken cancellationToken)
         => InteractAsync(@event, root => new Command.DeleteUser(root.Id), cancellationToken);
 }
